eliminated all of the warnings #306
Many were due to number formatting in strings while others were related
to pre-declaring methods which does not have to be done in modern
Objective-C and Xcode 5. I also removed ACEAutocomplete which does not
appear to be used in the app and changed MBProgressHUD to be used as a
CocoaPod reference. Now when warnings are discovered they can be
address immediately so they do not add up.

+++ SpotHopper/Libs/Categories/UIImage+FixOrientation.m
@@ -92,7 +92,7 @@
 - (NSData*)scaleToMaxSizeInKB:(NSInteger)sizeInKB {
     CGFloat compression = 0.9f;
     CGFloat maxCompression = 0.1f;
-    int maxFileSize = sizeInKB*1024;
+    NSInteger maxFileSize = sizeInKB*1024;
     
     NSData *imageData = UIImageJPEGRepresentation(self, compression);
     
@@ -101,7 +101,7 @@
         imageData = UIImageJPEGRepresentation(self, compression);
     }
     NSLog(@"Compression - %f", compression);
-    NSLog(@"Size - %d", [imageData length]);
+    NSLog(@"Size - %lu", [imageData length]);
     
     return imageData;
 }
